Don't exclude obj from artifacts
Currently we exclude the obj directory from the set of artifacts we
collect after a build as it contains only intermediate outputs rather
than the final outputs in the bin directory. However, those intermediate
outputs (such as the project.assets.json file) can be very useful in
diagnosing build problems.

Here we drop the obj directory exclusion so that we can get a copy of
the project.assets.json file from a PR build.