Remove /warnaserror
Our PR builds have been hit with https://github.com/dotnet/sdk/issues/35768,
where an issue in the template engine causes `dotnet new` to report an
error and return an exit code. We use MSBuild to script `dotnet new` to
create a solution to build on the fly, and that invocation of MSBuild
takes `/warnaserror` as an argument. When MSBuild reports a warning for
the exit code, that becomes an error that fails the whole build.

This issue is fixed in newer versions of the SDK, but present in the
version still installed on the PR build machine images. By removing
/warnaserror for the moment we hope to get past the issue in the SDK.

We will need to add /warnaserror back in later, once the build systems
have a version of the SDK with the fix. Otherwise we risk other warnings
being introduced without noticing.