Log when multiple up-to-date checks exist
CPS supports projects having multiple `IBuildUpToDateCheckProvider` implementations being active. When multiple exist, any one of them can declare the project as out of date.

This can cause confusion when there are multiple checks, as the log output from this check will declare the project up to date, then another check determines the project is out of date but doesn't log anything. In such cases, the output resembles:

```
1>FastUpToDate: Project is up-to-date. (MauiLib1)
1>------ Build started: Project: MauiLib1, Configuration: Debug Any CPU ------
1>MauiLib1 -> C:\Users\emafern\source\repos\MauiApp9\MauiLib1\bin\Debug\net8.0-ios\MauiLib1.dll
```

This change detects when multiple checks are present, and includes the following output:

```
1>FastUpToDate: Note that other checks may still determine the project is not up-to-date. The following checks are present: (MauiLib1)
1>FastUpToDate: - Microsoft.VisualStudio.OtherComponent.OtherCheck (MauiLib1)
```

This extra information avoids confusion and directs the reader to a specific implementation for further investigation.