Make code unit testable
The new use of `TryGetCurrentConfigurationGeneralPropertiesSnapshot` is incredibly painful to mock.

Furthermore, the old `SDKVersionTelemetryServiceComponent` was implemented as a multi-life component on a non-dynamic component.

We simplify things here by having only a single lifetime for the component, pulling the acquisition of the SDK version into a virtual method that can be mocked, and flattening two types into one (renaming it and the corresponding unit test in the process).

Also uses an xUnit theory to merge all tests into one.