Add a rule for AdditionalFiles
To ensure the language service can get the same link metadata and
handling as Compile items, I'm updating the VS Code project system to
reuse the same logic. This requires having a new rule to grab the items
during evaluation, so I'm copy/pasting the Compile item and updating
as appropriate.