Remove tests
The implementation changed considerably. There's not that much logic here to validate, but the code uses many layers of services and APIs that are a pain to mock. I don't think there's value in porting these tests, especially as we're no longer parsing strings. String parsing always requires a lot of tests, and we have removed that parsing from this PR.