Fix NU1701 warnings
Nuget raises the NU1701 warning when a referenced package does not contain an assembly that is compatible with the project's target framework. E.g., if the project targets .NET 7 but the package only contains assemblies targeting .NET Framework 4.7.2 then you'll get this warning. The restore will still complete but you may end up referencing an assembly that may fail to run correctly at runtime.

Early in our C# Dev Kit efforts we needed to turn off this warning because CPS didn't have .NET 7-compatible binaries yet and we _wanted_ to build against the available .NET Framework 4.7.2 binaries while we worked through our own multi-targeting plan. CPS has now had proper .NET 7/8 binaries for a while so we can turn the warning back on--but we also need to adjust our package references.

Currently we depend on the Microsoft.VisualStudio.ProjectSystem.SDK "metapackage" which brings in transitive references to the MS.VS.ProjectSystem, MS.VS.ProjectSystem.VS, and MS.VS.ProjectSystem.Sdk.Tools packages. The first has .NET 7/8 binaries available, and the third brings in build tools but no assemblies. The second is VS-specific and depends on .NET Framework 4.7.2 and it also depends on a bunch of other binaries that are also .NET Framework 4.7.2-specific. This is where all of our current NU1701 warnings would come from.

Here we remove the dependency on Microsoft.VisualStudio.ProjectSystem.SDK in favor of the specific packages we need; this allows us to turn the warning back on.

Note that this has benefits for consumers of our package as well. They would have gotten a transitive reference to the Microsoft.VisualStudio.ProjectSystem.VS package and had all the same NU1701 warnings, but would not have been in a position to fix it, just suppress it.