Use strongly-typed error reporter
Historically, the `IWorkspace` type was in the host-agnostic layer, meaning that the error-reporing type was specific to the host. Since the introduction of C# DevKit, `IWorkspace` moved into the VS layer, and we no longer need this property to be of type `object`. Instead, it can be typed as `IVsLanguageServiceBuildErrorReporter2`, which simplifies consuming code.