Use production OptProf tooling
We pick up certain pipeline templates providing OptProf functionality
from the DartLab.OptProf repository. Currently we get these from the
"main" branch of that repo, but as with other repos "main" is the
bleeding edge of OptProf changes. This potentially exposes us to bugs
and instability.

Instead we want to use the "Production" tag rather than the latest main.
This should be better tested and more stable.

This is based on
https://devdiv.visualstudio.com/DevDiv/_git/VS/pullrequest/515949.