Replace immutable dictionary with mutable dictionary
Traces show this code is contributing a fair amount to CPU during solution load. We should investigate this more deeply to understand if we can reduce how often this is called. For now, we can improve this a little by removing use of immutable dictionary here, as it's not needed.