The look-up table for intecepting providers can be shared, instead of being created on every InterceptProperties.
It is basically a MEF extension look-up table. Ideally, it could be shared among projects, but we can at least share
the dictionary within a project with little changes.

The problem is that the product would create one InterceptProperties for almost every property, just to read the value, and it throws away
the instance immediately, and create a new one for another property. Recreating the dicitonary becomes a heavy cost during solution loading time/