Don't NGEN for all architectures
Fixes AB#1987317.

We currently specify that Microsoft.VisualStudio.Editors.dll and
Microsoft.VisualStudio.AppDesigner.dll should be NGEN'd for all
architectures. This means that on an x64 system NGEN will run twice for
each: once to create a 32-bit version, and once to create a 64-bit
version. As we no longer ship a 32-bit version of VS the former is
unnecessary, and we're just wasting time and resources on that NGEN and
potentially delaying useful NGEN work.

This commit removes the `<NGenArchitecture>all</NGenArchitecture>`
MSBuild property affecting these binaries. Without this property being
set, NGEN will default to the current system architecture (e.g. AMD64 on
an AMD64 system).