Rename the legacy "Managed Resource Editor" editor factory
When a user right-clicks a `.resx` file in Solution Explorer and selects "Open With..." they are presented with a list of editor factories.

With the forthcoming introduction of the Resource Explorer, users are now presented with (among other options):

- "Resource Explorer"
- "Managed Resources Editor"

The second is the legacy experience. This change appends a "(legacy)" to the name of the editor to help differentiate between the two, and to help users find the improved editor.