Update CPS to version 11.17.207
Insertions from dotnet/project-system into dotnet-project-system-vscode automatically update the CPS version of that repo to match the version used in this repo. I need a newer version in dotnet-project-system-vscode for some feature work, so to avoid problems I'm updating it here as well.

This also allows us to remove a `PackageReference` that was needed just to resolve a Component Governance issue. The change also removes an unnecessary "using" that the build started to warn about.