Specify preferred persistence GUID for project types
This change takes a new version of CPS, and populates a new property on the `ProjectTypeRegistrationAttribute` that specifies the preferred project type GUID for persistence.

This helps avoid an issue where occasionally VS would change the project type GUID in the solution file, as now the project can explicitly define when it wants to use a different project type GUID for persistence, as we do for managed C#/VB/F# projects.