Fix lingering analyzer node with warning
Generally, we want MSBuild-based dependencies to exist in evaluation in order to be added to the tree. Items that exist only in build data are not displayed.

Analyzers opt out of this default behaviour as set `ResolvedItemRequiresEvaluatedItem` to false. Analyzers can be added by targets and should still be visible on screen.

When such an item is removed from build data, we only removed from the tree when no evaluation data was present.

There was a bug in this code. It was checking for an evaluated item in the previous data, where it should have been checking in the current data. It's a simple fix.